VER-80: some unstagged files.
