+++ src/components/shared.js
@@ -4,6 +4,7 @@ export const Button = styled('button')`
   height: 30px;
   border-radius: 8px;
   border: 1px solid #000;
+  padding: 0px 15px;
   background: #fff;
   margin-right: 3px;
   box-shadow: 0px 2px 0px 0px #000;
