+++ src/pages/Issue/fixtures.js
@@ -1,0 +1,23 @@
+export const checks = [
+    {
+      value: 'check-1',
+      label: 'temporary check 1',
+    },
+    {
+      value: 'check-2',
+      label: 'temporary check 2',
+    },
+    {
+      value: 'check-3',
+      label: 'temporary check 3',
+    },
+    {
+      value: 'check-4',
+      label: 'temporary check 4',
+    },
+    {
+      value: 'check-5',
+      label: 'temporary check 5',
+    },
+  ]
+
